# docs/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CreateApplicationRequest": {
        "additionalProperties": false,
        "description": "Request to create an application.",
        "properties": {
          "manifest": {
            "$ref": "#/components/schemas/ApplicationManifest",
            "description": "Application manifest to store and reconcile."
          }
        },
        "required": [
          "manifest"
        ],
        "type": "object"
      },
...
      "Envelope_AcceptedOperation": {
        "description": "Successful API response envelope.",
        "properties": {
          "data": {
            "description": "Operation accepted by an application mutation endpoint.",
            "properties": {
              "application_id": {
                "description": "Stable application identifier.",
                "type": "string"
              },
              "generation": {
                "description": "Generation created by the mutation.",
                "format": "int64",
                "minimum": 1,
                "type": "integer"
              },
              "operation_id": {
                "description": "Asynchronous operation identifier.",
                "type": "string"
              }
            },
            "required": [
              "operation_id",
              "application_id",
              "generation"
            ],
            "type": "object"
          }
        },
        "required": [
          "data"
        ],
        "type": "object"
      },
...
      "ErrorBody": {
        "description": "Structured error returned by the API.",
        "properties": {
          "code": {
            "description": "Stable machine-readable error code.",
            "type": "string"
          },
          "details": {
            "description": "Optional structured error details."
          },
          "message": {
            "description": "Safe human-readable error message.",
            "type": "string"
          },
          "request_id": {
            "description": "Server-generated request identifier.",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "piqueld control-plane API. Mutation responses identify durable operations; named volumes are retained on deletion.",
    "license": {
      "identifier": "Apache-2.0",
      "name": "Apache-2.0"
    },
    "title": "piqueld API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/applications": {
      "get": {
        "operationId": "listApplications",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Page_ApplicationView"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          }
        },
        "summary": "List applications"
      },
      "post": {
        "operationId": "createApplication",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 128,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateApplicationRequest"
              }
            },
            "application/toml": {
              "schema": {
                "type": "string"
              }
            },
            "text/toml": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_AcceptedOperation"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "502": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          }
        },
        "summary": "Create an application"
      }
    },
    "/api/v1/applications/plan": {
      "post": {
        "operationId": "planApplicationCreate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PlanApplicationRequest"
              }
            },
            "application/toml": {
              "schema": {
                "type": "string"
              }
            },
            "text/toml": {
              "schema": {
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_PlanView"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "415": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          }
        },
        "summary": "Preview application creation"
      }
    },
...
    "/api/v1/applications/{id}/status": {
      "get": {
        "operationId": "applicationStatus",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 8,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_ApplicationStatusView"
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Structured, sanitized error"
          }
        },
        "summary": "Get application status"
      }
    },
...
  },
  "servers": [
    {
      "description": "Default loopback TCP endpoint; clients may also use the configured Unix socket.",
      "url": "http://127.0.0.1:7845"
    }
  ]
}